producer_docker/requirements.txt
certifi==2019.6.16
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
kafka-python==1.4.6
MarkupSafe==1.1.1
numpy==1.16.4
opencv-python==4.1.0.25
pymongo==3.8.0
Werkzeug==0.15.5
